# Improve environment performance