@jdegoes 2024-05-12
@kyri-petrou Thanks for your work on this! The fast-path for Scope is clearly working.

I am concerned that there are a number of benchmarks that seem to get slower after the PR. Is this still a work-in-progress or are you looking for more final feedback here?