@kyri-petrou 2024-05-12
@jdegoes I'm looking for some final feedback; The work that it's left to do is adding tests getting MiMa happy. 

> I am concerned that there are a number of benchmarks that seem to get slower after the PR

There are only 2 benchmarks that become slower after this PR, [one that only benchmarks `add`](https://github.com/zio/zio/pull/8846/files#diff-cd203fbe9b8fe5eb32f9b55259ba090422746e1145b952b45968f3df20820263R44-R46) (old implementation was ~x2 faster)and another that has a [10:1 add-to-get ratio](https://github.com/zio/zio/pull/8846/files#diff-cd203fbe9b8fe5eb32f9b55259ba090422746e1145b952b45968f3df20820263R52-R72), where before it was only 20% faster.

On the other hand, [for a 1:10 add-to-get ratio](https://github.com/zio/zio/pull/8846/files#diff-cd203fbe9b8fe5eb32f9b55259ba090422746e1145b952b45968f3df20820263R97-R109)(which I think is a more representative usage pattern), we get ~10x improvement with the new implementation.

Previously, the `add` method was extremely simple (just an [`updated` on the immutable map](https://github.com/zio/zio/pull/8846/files#diff-4c59413d8710d8bcbf8186d9b6dc1575e8e5ffaea9f2df589c64860c74360306L160)), which meant that a lot of the complexity was delegated to `get`. Personally I think it's OK to sacrifice some write performance to get a much bigger improvement in reading performance, but at the same time I do understand your conservation. If you think the writting performance loss is too much, I can perhaps try optimizing it a bit better or look into alternative implementations