@jdegoes 2024-05-15
> This synthetic method was due to the parameter that had a default value. However, given that the constructor of ZEnvironment is private, I think it's relatively safe to add that as a MiMa exclusion?

@kyri-petrou Yes, for sure!